+++ src/screen/sys.rs
@@ -1,5 +1,0 @@
-#[cfg(unix)]
-pub(crate) mod unix;
-
-#[cfg(windows)]
-pub(crate) mod winapi;