+++ CHANGELOG.md
@@ -7,9 +7,12 @@
     documentation
 * Replace `docs/UPGRADE.md` with the [Upgrade Paths](https://github.com/crossterm-rs/crossterm/wiki#upgrade-paths)
   documentation
- - Rename `screen::Rawscreen::into_raw_mode` to `screen::Rawscreen::enable_raw_mode`
+ - Merged `screen` module into `terminal`  
  - Remove `screen::AlternateScreen`
- 
+ - Remove `screen::Rawscreen`
+    * Move and rename `Rawscreen::into_raw_mode` and `Rawscreen::disable_raw_mode` to `terminal::enable_raw_mode` and `terminal::disable_raw_mode`
+ - Move `screen::EnterAlternateScreen` and `screen::LeaveAlternateScreen` to `terminal::EnterAlternateScreen` and `terminal::LeaveAlternateScreen`
+  
 # Version 0.13.3
 
 - Remove thread from AsyncReader on Windows.